pyproject.toml
[project]
name = "agent-demo-galileo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.13.0",
    "llama-index-embeddings-openai>=0.5.0",
    "llama-index-llms-openai>=0.5.1",
    "llama-index-utils-workflow>=0.4.0",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "ragchecker>=0.1.9",
]
